+++ src/main/java/org/icgc/argo/program_service/converter/DataCenterConverter.java
@@ -18,25 +18,31 @@
  *
  */
 
-package org.icgc.argo.program_service.repositories;
+package org.icgc.argo.program_service.converter;
 
-import java.util.List;
-import java.util.Optional;
-import java.util.UUID;
-import org.icgc.argo.program_service.model.entity.ProgramEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
+import org.icgc.argo.program_service.model.dto.DataCenterDTO;
+import org.icgc.argo.program_service.model.dto.DataCenterRequestDTO;
+import org.icgc.argo.program_service.model.entity.DataCenterEntity;
+import org.mapstruct.Mapper;
+import org.mapstruct.Mapping;
+import org.mapstruct.MappingTarget;
 
-public interface ProgramRepository
-    extends JpaRepository<ProgramEntity, UUID>, JpaSpecificationExecutor<ProgramEntity> {
-  Optional<ProgramEntity> findByShortName(String name);
+@Mapper(
+    config = ConverterConfig.class,
+    uses = {CommonConverter.class})
+public interface DataCenterConverter {
+  DataCenterConverter INSTANCE = new DataCenterConverterImpl(CommonConverter.INSTANCE);
 
-  @Query("select distinct p.shortName from ProgramEntity p where p.active =true")
-  List<String> getActivePrograms();
+  DataCenterDTO dataCenterToDataCenterEntity(DataCenterEntity p);
 
-  @Query(
-      "SELECT P FROM ProgramEntity AS P INNER JOIN DataCenterEntity AS D ON P.dataCenterId=D.id WHERE D.shortName=:shortName AND P.active = false")
-  List<ProgramEntity> getActiveProgramsForDataCenter(@Param("shortName") String shortName);
+  @Mapping(target = "id", ignore = true)
+  @Mapping(target = "programEntities", ignore = true)
+  DataCenterEntity dataCenterToDataCenterEntity(DataCenterRequestDTO p);
+
+  @Mapping(target = "id", ignore = true)
+  @Mapping(target = "programEntities", ignore = true)
+  @Mapping(target = "shortName", ignore = true)
+  @Mapping(target = "submissionSongCode", ignore = true)
+  void updateDataCenter(
+      DataCenterEntity updatingDataCenter, @MappingTarget DataCenterEntity dataCenterToUpdate);
 }
