+++ src/main/java/org/icgc/argo/program_service/services/ProgramService.java
@@ -42,6 +42,7 @@ import javax.validation.ValidatorFactory;
 import lombok.NonNull;
 import lombok.extern.slf4j.Slf4j;
 import lombok.val;
+import org.icgc.argo.program_service.converter.DataCenterConverter;
 import org.icgc.argo.program_service.converter.ProgramConverter;
 import org.icgc.argo.program_service.model.dto.DataCenterRequestDTO;
 import org.icgc.argo.program_service.model.entity.*;
@@ -72,9 +73,11 @@ public class ProgramService {
   private final RegionRepository regionRepository;
   private final CountryRepository countryRepository;
   private final ProgramConverter programConverter;
+  private final DataCenterConverter dataCenterConverter;
   private final ProgramCancerRepository programCancerRepository;
   private final ProgramPrimarySiteRepository programPrimarySiteRepository;
   private final ProgramInstitutionRepository programInstitutionRepository;
+  private final ProgramRegionRepository programRegionRepository;
   private final ProgramCountryRepository programCountryRepository;
   private final ValidatorFactory validatorFactory;
 
@@ -85,12 +88,14 @@ public class ProgramService {
       @NonNull CancerRepository cancerRepository,
       @NonNull PrimarySiteRepository primarySiteRepository,
       @NonNull ProgramConverter programConverter,
+      @NonNull DataCenterConverter dataCenterConverter,
       @NonNull ProgramCancerRepository programCancerRepository,
       @NonNull ProgramPrimarySiteRepository programPrimarySiteRepository,
       @NonNull InstitutionRepository institutionRepository,
       @NonNull RegionRepository regionRepository,
       @NonNull CountryRepository countryRepository,
       @NonNull ProgramInstitutionRepository programInstitutionRepository,
+      @NonNull ProgramRegionRepository programRegionRepository,
       @NonNull ProgramCountryRepository programCountryRepository,
       @NonNull ValidatorFactory validatorFactory) {
     this.programRepository = programRepository;
@@ -98,12 +103,14 @@ public class ProgramService {
     this.cancerRepository = cancerRepository;
     this.primarySiteRepository = primarySiteRepository;
     this.programConverter = programConverter;
+    this.dataCenterConverter = dataCenterConverter;
     this.programCancerRepository = programCancerRepository;
     this.programPrimarySiteRepository = programPrimarySiteRepository;
     this.institutionRepository = institutionRepository;
     this.regionRepository = regionRepository;
     this.countryRepository = countryRepository;
     this.programInstitutionRepository = programInstitutionRepository;
+    this.programRegionRepository = programRegionRepository;
     this.programCountryRepository = programCountryRepository;
     this.validatorFactory = validatorFactory;
   }
@@ -414,28 +421,30 @@ public class ProgramService {
   }
 
   public DataCenterEntity createDataCenter(DataCenterRequestDTO dataCenterRequestDTO) {
-    val dataCenterEntity = programConverter.dataCenterToDataCenterEntity(dataCenterRequestDTO);
+    val dataCenterEntity = dataCenterConverter.dataCenterToDataCenterEntity(dataCenterRequestDTO);
     val d = dataCenterRepository.save(dataCenterEntity);
     return d;
   }
+
   public DataCenterEntity updateDataCenter(
-          @NonNull DataCenterEntity dataCenterToUpdate, @NonNull DataCenterEntity updatingDataCenter) {
-    programConverter.updateDataCenter(updatingDataCenter, dataCenterToUpdate);
+      @NonNull DataCenterEntity dataCenterToUpdate, @NonNull DataCenterEntity updatingDataCenter) {
+    dataCenterConverter.updateDataCenter(updatingDataCenter, dataCenterToUpdate);
     dataCenterRepository.save(dataCenterToUpdate);
     return dataCenterToUpdate;
   }
 
   public DataCenterEntity findDataCenterByShortName(@NonNull String name) {
     val search =
-            dataCenterRepository.findOne(new DataCenterSpecificationBuilder().buildByShortName(name));
+        dataCenterRepository.findOne(new DataCenterSpecificationBuilder().buildByShortName(name));
 
     if (search.isEmpty()) {
       throw Status.NOT_FOUND
-              .withDescription("DataCenter '" + name + "' not found")
-              .asRuntimeException();
+          .withDescription("DataCenter '" + name + "' not found")
+          .asRuntimeException();
     }
     return search.get();
   }
+
   public List<String> getAllProgramNames() {
     val programNames = programRepository.getActivePrograms();
     return programNames;
@@ -493,5 +502,11 @@ public class ProgramService {
       ProgramEntity program, Set<CountryEntity> countries) {
     val id = program.getId();
     return c -> c.getProgram().getId().equals(id) && countries.contains(c.getCountry());
+  }
+
+  private static Predicate<ProgramRegion> programRegionPredicate(
+      ProgramEntity program, Set<RegionEntity> regions) {
+    val id = program.getId();
+    return r -> r.getProgram().getId().equals(id) && regions.contains(r.getRegion());
   }
 }
