+++ src/main/java/org/icgc/argo/program_service/services/ValidationService.java
@@ -181,4 +181,8 @@ public class ValidationService {
   public Set<String> validCountries() {
     return mapToSet(programService.listCountries(), CountryEntity::getName);
   }
+
+  public Set<String> validRegions() {
+    return mapToSet(programService.listRegions(), RegionEntity::getName);
+  }
 }
